Refactor global-setup.ts
